@@ -1,89 +1,73 @@
 import { Routes } from '@angular/router';
+import { HomeComponent } from './shared/home/home.component';
 import { LoginComponent } from './pages/login/login.component';
 import { RegisterComponent } from './pages/register/register.component';
 import { AppointmentsComponent } from './pages/appointments/appointments.component';
 import { DoctorsComponent } from './pages/doctors/doctors.component';
 import { AuthGuard } from './guards/auth.guard';
-import { AdminLayoutComponent } from './layouts/admin-layout/admin-layout.component';
-import { DoctorLayoutComponent } from './layouts/doctor-layout/doctor-layout.component';
+import { AdminGuard } from './guards/admin.guard';
+import { ApiTesterComponent } from '../app/pages/api-tester/api-tester.component';
+import { AdminDashboardComponent } from './admin/components/dashboard/admin-dashboard.component';
 import { DoctorDashboardComponent } from './pages/doctors/doctor-dashboard/doctor-dashboard.component';
-import { DoctorAppointmentsComponent } from './pages/doctors/doctor-appointments/doctor-appointments.component';
-import { adminGuard } from './guards/admin.guard';
+import { PatientDashboardComponent } from './pages/patient/patient-dashboard/patient-dashboard.component';
+
 
 export const routes: Routes = [
-  { path: '', redirectTo: '/login', pathMatch: 'full' },
+
+  { path: 'api-tester', component: ApiTesterComponent },
+  
+  // ✅ PÁGINA PRINCIPAL (primera que se ve)
+  { path: '', component: HomeComponent },
+  
+  // Rutas de autenticación
   { path: 'login', component: LoginComponent },
   { path: 'register', component: RegisterComponent },
-  { path: 'appointments', component: AppointmentsComponent, canActivate: [AuthGuard] },
+  
+  // Rutas públicas
+  { path: 'appointments', component: AppointmentsComponent },
   { path: 'doctors', component: DoctorsComponent },
-
-  // Rutas de Admin
+  
+  // Rutas protegidas de Admin
   {
     path: 'admin',
-    component: AdminLayoutComponent,
-    canActivate: [adminGuard],
-    children: [
-      {
-        path: '',
-        loadComponent: () => import('./admin/components/dashboard/admin-dashboard.component')
-          .then(m => m.AdminDashboardComponent)
-      },
-      {
-        path: 'dashboard',
-        loadComponent: () => import('./admin/components/dashboard/admin-dashboard.component')
-          .then(m => m.AdminDashboardComponent)
-      },
-      {
-        path: 'users',
-        loadComponent: () => import('./admin/components/users/users-management.component')
-          .then(m => m.UsersManagementComponent)
-      },
-      {
-        path: 'appointments',
-        loadComponent: () => import('./admin/components/appointments/appointments-management.component')
-          .then(m => m.AppointmentsManagementComponent)
-      },
-    ]
+    canActivate: [AuthGuard, AdminGuard],
+    loadChildren: () => import('./admin/admin.routes').then(m => m.adminRoutes)
   },
-
-  // Rutas de Paciente
+  
+  // Rutas protegidas de Paciente
   {
     path: 'patient',
     canActivate: [AuthGuard],
-    children: [
-      {
-        path: '',
-        loadComponent: () => import('./pages/patient/patient-dashboard/patient-dashboard.component')
-          .then(m => m.PatientDashboardComponent)
-      },
-      {
-        path: 'dashboard',
-        loadComponent: () => import('./pages/patient/patient-dashboard/patient-dashboard.component')
-          .then(m => m.PatientDashboardComponent)
-      }
-    ]
+    loadChildren: () => import('./pages/patient/patient.routes').then(m => m.patientRoutes)
   },
-
-  // Rutas de Doctor
+  
+  // Rutas protegidas de Doctor
   {
     path: 'doctor',
-    component: DoctorLayoutComponent,
     canActivate: [AuthGuard],
-    children: [
-      {
-        path: '',
-        component: DoctorDashboardComponent
-      },
-      {
-        path: 'dashboard',
-        component: DoctorDashboardComponent
-      },
-      {
-        path: 'appointments',
-        component: DoctorAppointmentsComponent
-      },
-    ]
+    loadChildren: () => import('./pages/doctors/doctor.routes').then(m => m.doctorRoutes)
   },
+  
+    { 
+    path: 'admin/dashboard', 
+    component: AdminDashboardComponent,
+    canActivate: [AuthGuard]
+  },
+  { 
+    path: 'doctor/dashboard', 
+    component: DoctorDashboardComponent,
+    canActivate: [AuthGuard]
+  },
+  { 
+    path: 'patient/dashboard', 
+    component: PatientDashboardComponent,
+    canActivate: [AuthGuard]
+  },
+  
+  // Otras rutas...
 
-  { path: '**', redirectTo: '/login' }
+  // Redirigir rutas no encontradas a home
+  { path: '**', 
+    redirectTo: '' },
+  
 ];
