@@ -1,38 +1,35 @@
 import { Injectable } from '@angular/core';
-import { 
-  CanActivate, 
-  ActivatedRouteSnapshot,
-  RouterStateSnapshot,
-  Router
-} from '@angular/router';
+import { CanActivate, Router } from '@angular/router';
+import { Observable } from 'rxjs';
+import { map, take } from 'rxjs/operators';
 import { AuthService } from '../services/auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthGuard implements CanActivate {
-  
+export class AdminGuard implements CanActivate {
   constructor(
     private authService: AuthService,
     private router: Router
   ) {}
-  
-  canActivate(
-    route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot
-  ): boolean {
-    const requiredRole = route.data['role'] as string;
-    
-    if (!this.authService.isLoggedIn()) {
-      this.router.navigate(['/login']);
-      return false;
-    }
-    
-    if (requiredRole && !this.authService.hasRole(requiredRole)) {
-      this.router.navigate(['/']);
-      return false;
-    }
-    
-    return true;
+
+  canActivate(): Observable<boolean> {
+    return this.authService.currentUser$.pipe(
+      take(1),
+      map(user => {
+        if (!user || user.role !== 'admin') {
+          this.router.navigate(['/']);
+          return false;
+        }
+        return true;
+      })
+    );
   }
 }
+
+import { CanActivateFn } from '@angular/router';
+
+export const AuthGuard: CanActivateFn = (route, state) => {
+  // tu lógica aquí
+  return true;
+};
