@@ -1,86 +1,177 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { BehaviorSubject, Observable, tap } from 'rxjs';
-
-export interface MedicalAppointment {
-  id: number;
-  patientId: number;
-  patientName: string;
-  patientAge?: number;
-  doctorId: number;
-  doctorName: string;
-  specialty: string;
-  date: string;
-  time: string;
-  endTime?: string;
-  status: 'scheduled' | 'confirmed' | 'in-progress' | 'completed' | 'cancelled' | 'emergency' | 'rescheduled';
-  priority?: 'low' | 'medium' | 'high' | 'urgent';
-  reason: string;
-  room?: string;
-  location?: string;
-  notes?: string;
-  createdAt: Date;
-  previousDate?: string;
-  previousTime?: string;
-}
+import { Observable, BehaviorSubject } from 'rxjs';
+import { tap, map } from 'rxjs/operators';
+import { ApiService, Appointment } from './api.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AppointmentsService {
-  private apiUrl = 'http://localhost:8000/api/appointments/';
-  private appointmentsSubject = new BehaviorSubject<MedicalAppointment[]>([]);
+  private appointmentsSubject = new BehaviorSubject<Appointment[]>([]);
+  public appointments$ = this.appointmentsSubject.asObservable();
 
-  constructor(private http: HttpClient) {
+  constructor(private apiService: ApiService) {
     this.loadAppointments();
   }
 
-  private loadAppointments(): void {
-    this.http.get<MedicalAppointment[]>(this.apiUrl).subscribe(appointments => {
-      this.appointmentsSubject.next(appointments);
+  // **CARGAR CITAS**
+  loadAppointments(): void {
+    this.apiService.getAppointments().subscribe({
+      next: (appointments) => {
+        this.appointmentsSubject.next(appointments);
+      },
+      error: (error) => {
+        console.error('Error cargando citas:', error);
+      }
     });
   }
 
-  getAllAppointments(): Observable<MedicalAppointment[]> {
-    return this.appointmentsSubject.asObservable();
+  // **OBTENER CITAS**
+  getAllAppointments(): Observable<Appointment[]> {
+    return this.appointments$;
   }
 
-  createAppointment(appointmentData: Omit<MedicalAppointment, 'id' | 'createdAt'>): Observable<MedicalAppointment> {
-    return this.http.post<MedicalAppointment>(this.apiUrl, appointmentData).pipe(
+  getAppointment(id: number): Observable<Appointment> {
+    return this.apiService.getAppointment(id);
+  }
+
+  getPatientAppointments(patientId: number): Observable<Appointment[]> {
+    return this.appointments$.pipe(
+      map(appointments => appointments.filter(apt => apt.patient_id === patientId))
+    );
+  }
+
+  getDoctorAppointments(doctorId: number): Observable<Appointment[]> {
+    return this.appointments$.pipe(
+      map(appointments => appointments.filter(apt => apt.doctor_id === doctorId))
+    );
+  }
+
+  getTodayAppointments(): Observable<Appointment[]> {
+    const today = new Date().toISOString().split('T')[0];
+    return this.appointments$.pipe(
+      map(appointments => appointments.filter(apt => apt.date === today))
+    );
+  }
+
+  getUpcomingAppointments(): Observable<Appointment[]> {
+    const today = new Date().toISOString().split('T')[0];
+    return this.appointments$.pipe(
+      map(appointments => appointments.filter(apt => 
+        apt.date >= today && (apt.status === 'scheduled' || apt.status === 'confirmed')
+      ))
+    );
+  }
+
+  // **CREAR CITA**
+  createAppointment(appointmentData: Partial<Appointment>): Observable<Appointment> {
+    return this.apiService.createAppointment(appointmentData).pipe(
       tap(() => this.loadAppointments())
     );
   }
 
-  updateAppointment(updatedData: Partial<MedicalAppointment>): Observable<MedicalAppointment> {
-    return this.http.patch<MedicalAppointment>(`${this.apiUrl}${updatedData.id}/`, updatedData).pipe(
+  // **ACTUALIZAR CITA**
+  updateAppointment(id: number, appointmentData: Partial<Appointment>): Observable<Appointment> {
+    return this.apiService.updateAppointment(id, appointmentData).pipe(
       tap(() => this.loadAppointments())
     );
   }
 
-  updateAppointmentStatus(id: number, status: MedicalAppointment['status']): Observable<MedicalAppointment> {
-    return this.updateAppointment({ id, status });
+  // **OPERACIONES ESPECÍFICAS**
+  confirmAppointment(id: number): Observable<Appointment> {
+    return this.updateAppointment(id, { status: 'confirmed' });
   }
 
+  cancelAppointment(id: number, reason?: string): Observable<Appointment> {
+    const updateData: Partial<Appointment> = { 
+      status: 'cancelled'
+    };
+    
+    if (reason) {
+      updateData.notes = reason;
+    }
+    
+    return this.updateAppointment(id, updateData);
+  }
+
+  startAppointment(id: number): Observable<Appointment> {
+    return this.updateAppointment(id, { status: 'in-progress' });
+  }
+
+  completeAppointment(id: number, notes?: string): Observable<Appointment> {
+    const updateData: Partial<Appointment> = { 
+      status: 'completed'
+    };
+    
+    if (notes) {
+      updateData.notes = notes;
+    }
+    
+    return this.updateAppointment(id, updateData);
+  }
+
+  rescheduleAppointment(id: number, newDate: string, newTime: string): Observable<Appointment> {
+    return this.updateAppointment(id, { 
+      date: newDate, 
+      time: newTime,
+      status: 'scheduled'
+    });
+  }
+
+  // **ELIMINAR CITA (solo admin)**
   deleteAppointment(id: number): Observable<any> {
-    return this.http.delete(`${this.apiUrl}${id}/`).pipe(
+    return this.apiService.deleteAppointment(id).pipe(
       tap(() => this.loadAppointments())
     );
   }
 
-  rescheduleAppointment(id: number, newDate: string, newTime: string): Observable<MedicalAppointment> {
-    return this.http.patch<MedicalAppointment>(`${this.apiUrl}${id}/reschedule/`, {
-      date: newDate,
-      time: newTime
-    }).pipe(
-      tap(() => this.loadAppointments())
+  // **ESTADÍSTICAS**
+  getAppointmentStats(): Observable<any> {
+    return this.appointments$.pipe(
+      map(appointments => {
+        const today = new Date().toISOString().split('T')[0];
+        
+        return {
+          total: appointments.length,
+          today: appointments.filter(apt => apt.date === today).length,
+          pending: appointments.filter(apt => apt.status === 'scheduled').length,
+          completed: appointments.filter(apt => apt.status === 'completed').length,
+          cancelled: appointments.filter(apt => apt.status === 'cancelled').length
+        };
+      })
     );
   }
 
-  getAppointmentsByDoctor(doctorId: number): Observable<MedicalAppointment[]> {
-    return this.http.get<MedicalAppointment[]>(`${this.apiUrl}?doctor=${doctorId}`);
+  // **UTILIDADES**
+  getStatusText(status: string): string {
+    switch (status) {
+      case 'scheduled': return 'Programada';
+      case 'confirmed': return 'Confirmada';
+      case 'in-progress': return 'En Progreso';
+      case 'completed': return 'Completada';
+      case 'cancelled': return 'Cancelada';
+      default: return status;
+    }
   }
 
-  getAppointmentsByPatient(patientId: number): Observable<MedicalAppointment[]> {
-    return this.http.get<MedicalAppointment[]>(`${this.apiUrl}?patient=${patientId}`);
+  getStatusColor(status: string): string {
+    switch (status) {
+      case 'scheduled': return '#ffa500';
+      case 'confirmed': return '#28a745';
+      case 'in-progress': return '#007bff';
+      case 'completed': return '#6f42c1';
+      case 'cancelled': return '#dc3545';
+      default: return '#6c757d';
+    }
+  }
+
+  getPriorityColor(priority: string): string {
+    switch (priority) {
+      case 'urgent': return '#dc3545';
+      case 'high': return '#fd7e14';
+      case 'medium': return '#ffc107';
+      case 'low': return '#28a745';
+      default: return '#6c757d';
+    }
   }
 }
