@@ -1,9 +1,8 @@
-export interface ExtendedUser {
+export interface User {
   id: number | string;
   name: string;
   email: string;
   role: 'patient' | 'doctor' | 'admin';
-  profile?: any;
   phone?: string;
   birthdate?: string;
   gender?: 'M' | 'F' | 'O';
@@ -13,23 +12,22 @@ export interface ExtendedUser {
   emergencyContact?: string;
   createdAt?: string;
   specialty?: string;
-  dateOfBirth?: string; // Para compatibilidad
 }
 
-export interface PatientProfile extends ExtendedUser {
+export interface PatientProfile extends User {
   role: 'patient';
   medicalHistory?: string;
   priority?: 'low' | 'medium' | 'high' | 'urgent';
 }
 
-export interface DoctorProfile extends ExtendedUser {
+export interface DoctorProfile extends User {
   role: 'doctor';
   specialty: string;
   licenseNumber?: string;
   experienceYears?: number;
 }
 
-export interface AdminProfile extends ExtendedUser {
+export interface AdminProfile extends User {
   role: 'admin';
   permissions?: string[];
   lastLogin?: string;
